+++ Domain/EntityRequest/Ticket/TicketCreateRequest.cs
@@ -10,7 +10,7 @@ public class TicketCreateRequest
     [Required] public DateTime CreateDate { get; set; }
 
     public DateTime? SolveDate { get; set; }
-    public decimal? Amount { get; set; }
+    public double? Amount { get; set; }
     public string Status { get; set; } = null!;
     [Required] public int TicketTypeId { get; set; }
 }
